Fix compilation with recent rustc.
This wasn't building with rustc 1.37 for (at least) the following
reasons:

- Use of removed, previously-unstable features
- Linker args incompatible with new default linker (lld).

Updating the deps to versions that would build on current nightly
required some name changes etc.

This code *almost* builds on stable -- the remaining issues are an
ancient dependency relying on unstable versions of later-stabilized
features (linked-list-allocator), and the use of the OOM lang item in
main.